* fix some int warning

+++ IMiPhone/IMNWProxy.m
@@ -36,7 +36,7 @@ static IMNWProxy *sharedNWProxy = nil;
 
 - (NSError *)processErrorCode:(NSInteger)errorcode fromSource:(NSString *)source
 {
-    NSAssert(YES, @"Network connect response error: %li", errorcode);
+    NSAssert(YES, @"Network connect response error: %li", (long)errorcode);
     NSNumber *errorCodeNumber = [NSNumber numberWithLong:errorcode];
     NSString *errorMessage = [errorCodeNumber errorMessage];
     NSDictionary *userInfo = [NSDictionary dictionaryWithObject:errorMessage
